Move elicitation to its own folder in tests and factor out schema projection and validation tests to new file
